+++ apps/ledger-live-mobile/src/components/RootNavigator/types/LandingPagesNavigator.ts
@@ -1,7 +1,11 @@
+import { KeysPriceChange } from "@ledgerhq/live-common/market/utils/types";
 import { ScreenName } from "~/const";
 import { LandingPageUseCase } from "~/dynamicContent/types";
 
 export type LandingPagesNavigatorParamList = {
   [ScreenName.GenericLandingPage]: { useCase: LandingPageUseCase };
-  [ScreenName.LargeMoverLandingPage]: undefined;
+  [ScreenName.LargeMoverLandingPage]: {
+    currencyIds: string[];
+    initialRange: KeysPriceChange;
+  };
 };