+++ e2e/mobile/helpers/commonHelpers.ts
@@ -1,4 +1,4 @@
-import { findFreePort, close as closeBridge, init as initBridge } from "../bridge/server";
+import { close as closeBridge, findFreePort, init as initBridge } from "../bridge/server";
 import { getEnv, setEnv } from "@ledgerhq/live-env";
 import { exec } from "child_process";
 import { device, log } from "detox";
@@ -6,14 +6,7 @@ import { allure } from "jest-allure2-reporter/api";
 
 const BASE_DEEPLINK = "ledgerlive://";
 
-export const itifAndroid = (...args: Parameters<typeof test>) =>
-  isAndroid() ? test(...args) : test.skip("[Android only] " + args[0], args[1], args[2]);
-export const describeifAndroid = (...args: Parameters<typeof describe>) =>
-  isAndroid() ? describe(...args) : describe.skip("[Android only] " + args[0], args[1]);
 export const currencyParam = "?currency=";
-export const recipientParam = "&recipient=";
-export const amountParam = "&amount=";
-export const accountIdParam = "?accountId=";
 
 /**
  * Waits for a specified amount of time
@@ -55,7 +48,7 @@ export async function launchApp() {
       locale: "en-US",
     },
     permissions: {
-      camera: "YES", // Give iOS permissions for the camera
+      camera: "YES",
     },
   });
   return port;
@@ -78,14 +71,15 @@ export function setupEnvironment() {
   }
 }
 
-export const logMemoryUsage = async () => {
+export const logMemoryUsage = async (): Promise<void> => {
   const pid = process.pid;
   const isLinux = process.platform !== "darwin";
+  const topArgs = isLinux ? `-b -n 1 -p ${pid}` : `-l 1 -pid ${pid}`;
   exec(
-    `top ${isLinux ? "-b -n 1 -p" : "-l 1 -pid"} ${pid} | grep "${pid}" | awk '{print ${isLinux ? "$6" : "$8"}}'`,
-    async (error, stdout, stderr) => {
+    `top ${topArgs} | grep "${pid}" | awk '{print ${isLinux ? "$6" : "$8"}}'`,
+    async (error: Error | null, stdout: string, stderr: string): Promise<void> => {
       if (error || stderr) {
-        console.error(`Error getting memory usage:\n Error: ${error}\n Stderr: ${stderr}`);
+        log.error(`Error getting memory usage:\n Error: ${error}\n Stderr: ${stderr}`);
         return;
       }
       const logMessage = `📦 Detox Memory Usage: ${stdout.trim()}`;