+++ e2e/mobile/models/currencies.ts
@@ -1,0 +1,4 @@
+import { findCryptoCurrencyById } from "@ledgerhq/live-common/currencies/index";
+
+export const getCurrencyManagerApp = (currencyId: string) =>
+  findCryptoCurrencyById(currencyId)?.managerAppName.toLowerCase().split(" ")[0];