+++ libs/coin-modules/coin-aptos/src/__tests__/network/client.test.ts
@@ -86,7 +86,7 @@ describe("Aptos API", () => {
   });
 
   describe("getAccountInfo", () => {
-    it("calls getBalance, fetchTransactions and getHeight", async () => {
+    it("calls getCoinBalance, fetchTransactions and getHeight", async () => {
       mockedAptos.mockImplementation(() => ({
         view: jest.fn().mockReturnValue(["123"]),
         getTransactionByVersion: jest.fn().mockReturnValue({
@@ -126,7 +126,7 @@ describe("Aptos API", () => {
       mockedApolloClient.mockImplementation(() => ({
         query: async () => ({
           data: {
-            address_version_from_move_resources: [{ transaction_version: "v1" }],
+            account_transactions: [{ transaction_version: 1 }],
           },
           loading: false,
           networkStatus: 7,
@@ -192,7 +192,7 @@ describe("Aptos API", () => {
       mockedApolloClient.mockImplementation(() => ({
         query: async () => ({
           data: {
-            address_version_from_move_resources: [{ transaction_version: "v1" }],
+            account_transactions: [{ transaction_version: 1 }],
           },
           loading: false,
           networkStatus: 7,
@@ -256,7 +256,7 @@ describe("Aptos API", () => {
       mockedApolloClient.mockImplementation(() => ({
         query: async () => ({
           data: {
-            address_version_from_move_resources: [{ transaction_version: "v1" }],
+            account_transactions: [{ transaction_version: 1 }],
           },
           loading: false,
           networkStatus: 7,
@@ -310,7 +310,7 @@ describe("Aptos API", () => {
       mockedApolloClient.mockImplementation(() => ({
         query: async () => ({
           data: {
-            address_version_from_move_resources: [{ transaction_version: "v1" }],
+            account_transactions: [{ transaction_version: 1 }],
           },
           loading: false,
           networkStatus: 7,
@@ -476,7 +476,6 @@ describe("Aptos API", () => {
       });
     });
   });
-
   describe("broadcast", () => {
     it("broadcasts the transaction", async () => {
       mockedPost.mockImplementation(async () => ({ data: { hash: "ok" } }));