+++ libs/coin-modules/coin-solana/src/types.ts
@@ -13,6 +13,7 @@ import {
 import { ValidatorsAppValidator } from "./network/validator-app";
 import { TokenAccountState } from "./network/chain/account/token";
 import { PARSED_PROGRAMS } from "./network/chain/program/constants";
+import { SPLToken } from "@ledgerhq/cryptoassets/data/spl";
 
 export type TransferCommand = {
   kind: "transfer";
@@ -228,6 +229,7 @@ export type TransactionModel = { commandDescriptor?: CommandDescriptor } & (
 export type Transaction = TransactionCommon & {
   family: "solana";
   model: TransactionModel;
+  raw?: string;
 };
 
 export type TransactionRaw = TransactionCommonRaw & {
@@ -291,6 +293,7 @@ export type SolanaPreloadDataV1 = {
   version: "1";
   validatorsWithMeta: SolanaValidatorWithMeta[];
   validators: ValidatorsAppValidator[];
+  splTokens: SPLToken[] | null;
 };
 
 // exists for discriminated union to work