+++ libs/ledgerjs/packages/swift-bridge-hw-app-eth/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ledgerhq/swift-bridge-hw-app-eth
 
+## 1.2.31
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-app-eth@6.45.4
+
+## 1.2.31-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-app-eth@6.45.4-next.0
+
 ## 1.2.30
 
 ### Patch Changes